+++ lib/src/ui/components/boxes/questionnaire_date_time_item.dart
@@ -58,7 +58,8 @@ class QuestionnaireDateTimeItem extends QuestionnaireBaseItem
           children: [
             Padding(
               padding: const EdgeInsets.only(right: 5),
-              child: getItemIcon(questionnaireItem.type),
+              child: getItemIcon(questionnaireItem.type,
+                  Theme.of(context).colorScheme.primary),
             ),
             Text(displayText),
           ],